+++ packages/extensions/core/package.json
@@ -41,6 +41,7 @@
   },
   "typings": "./dist/exports.d.ts",
   "devDependencies": {
+    "@autorest/configuration": "~1.0.0",
     "@azure-tools/async-io": "~3.0.0",
     "@azure-tools/codegen": "~2.5.293",
     "@azure-tools/datastore": "~4.1.271",
