+++ packages/extensions/core/src/lib/configuration/index.ts
@@ -1,7 +1,6 @@
 export * from "./configuration-loader";
-export * from "./autorest-raw-configuration";
+export * from "./configuration-merging";
 export * from "./autorest-configuration";
 export * from "./autorest-context";
 export * from "./message-emitter";
-export * from "./directive";
 export * from "./configuration-file-resolver";
