+++ packages/libs/configuration/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./autorest-configuration";
+export * from "./autorest-raw-configuration";
+export * from "./directive";
